cms_bot/requirements.txt
openai
python-dotenv
Pillow 
pandas
crewai
spacy
langchain
'crewai[tools]'